requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
altair==5.4.1
    # via
    #   -r requirements.in
    #   streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via httpx
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via streamlit
contourpy==1.3.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
exceptiongroup==1.2.2
    # via anyio
fastcore==1.7.19
    # via nixtla
fonttools==4.54.1
    # via matplotlib
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.6
    # via httpx
httpx==0.27.2
    # via nixtla
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
kiwisolver==1.4.7
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.9.2
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.10.0
    # via altair
nixtla==0.6.1
    # via -r requirements.in
numpy==2.1.2
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
    #   utilsforecast
orjson==3.10.10
    # via nixtla
packaging==24.1
    # via
    #   altair
    #   fastcore
    #   matplotlib
    #   streamlit
    #   utilsforecast
pandas==2.2.3
    # via
    #   -r requirements.in
    #   nixtla
    #   streamlit
    #   utilsforecast
pillow==10.4.0
    # via
    #   matplotlib
    #   streamlit
protobuf==5.28.2
    # via streamlit
pyarrow==17.0.0
    # via streamlit
pydantic==2.9.2
    # via nixtla
pydantic-core==2.23.4
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   streamlit
rich==13.9.3
    # via streamlit
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
streamlit==1.39.0
    # via -r requirements.in
tenacity==9.0.0
    # via
    #   nixtla
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.1
    # via streamlit
tqdm==4.66.5
    # via nixtla
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   pydantic
    #   pydantic-core
    #   rich
    #   streamlit
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
utilsforecast==0.2.5
    # via nixtla
watchdog==5.0.3
    # via streamlit
